requirements.txt
numpy
matplotlib
opencv-python
Pillow
pandas
sns
scikit-learn